Navbar responsiveness and design fixes
Implemented responsiveness to the navbar and fixed design errors
