+++ app/Lucky_Folder/Dashboard/Dashboard.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import NavbarComponent from '../Components/UI/Navbar/NavbarComponent'
+import HeaderComponent from '../Components/UI/Header/HeaderComponent'
+import FriendsListComponent from '../Components/UI/FriendsList/FriendsListComponent'
+
+
+export default function page() {
+  return (
+    <div className='bg-[#242526] h-screen'>
+        <NavbarComponent/>
+        <HeaderComponent/>
+        <FriendsListComponent/>
+    </div>
+  )
+}
