+++ app/Lucky_Folder/page.tsx
@@ -1,15 +1,11 @@
 import React from 'react'
-import NavbarComponent from './Components/Navbar/NavbarComponent'
-import HeaderComponent from './Components/Header/HeaderComponent'
-import FriendsListComponent from './Components/FriendsList/FriendsListComponent'
+import Dashboard from '../Lucky_Folder/Dashboard/Dashboard'
 
 
 export default function page() {
   return (
-    <div className='bg-[#242526] h-screen'>
-        <NavbarComponent/>
-        <HeaderComponent/>
-        <FriendsListComponent/>
-    </div>
+    <>
+    <Dashboard/>
+    </>
   )
 }
